findemy-website/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
mysql-connector-python==8.1.0
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2
blinker==1.6.2
gunicorn==21.2.0
